Thêm Crystal Report + Chỉnh sửa SQL (Chơi mà để 1 SP/DV trên 1 MaHD gì kỳ v mấy 3)

+++ Doan_NET/ViewModel/User_DichVu_VM.cs
@@ -92,7 +92,7 @@ namespace Doan_NET.ViewModel
                 {
                     MaMatHang = dichVu.MaPT,
                     TenMatHang = dichVu.Ten,
-                    DonGia = (int)(dichVu.Gia ?? 0),
+                    DonGia = (decimal)(dichVu.Gia ?? 0),
                     SoLuong = 1
                 });
             }
